tox-requirements.txt
django-anymail
django-pytest
django-render-block
html2text
mock
pytest
pytest-django
pytest-pythonpath
tox
